Moved Coding and CodeableConcept to their own modules Also added IETF_LANGUAGE_TAG and LocaleConstants for consistent use.
