+++ portal/models/extension.py
@@ -2,7 +2,8 @@
 from abc import ABCMeta, abstractmethod, abstractproperty
 from flask import abort
 import pytz
-from sqlalchemy.orm.exc import NoResultFound
+
+from .coding import Coding
 
 
 class Extension:
@@ -35,8 +36,6 @@ class CCExtension(Extension):
             }
 
     def apply_fhir(self):
-        from .fhir import Coding  # local due to cycle
-
         assert self.extension['url'] == self.extension_url
         # track current concepts - must remove any not requested
         remove_if_not_requested = {e.code: e for e in self.children}
