+++ portal/models/intervention_strategies.py
@@ -21,9 +21,11 @@ from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm.exc import NoResultFound, MultipleResultsFound
 import sys
 
+from .codeable_concept import CodeableConcept
+from .coding import Coding
 from ..database import db
 from ..date_tools import localize_datetime
-from .fhir import CC, Coding, CodeableConcept
+from .fhir import CC
 from .identifier import Identifier
 from .intervention import Intervention, INTERVENTION, UserIntervention
 from .organization import Organization, OrgTree, OrganizationIdentifier
